+++ src/felupe/constitution/strain/models/__init__.py
@@ -3,9 +3,11 @@ from ._linear_elastic_plastic_isotropic import (
     LinearElasticPlasticIsotropicHardening,
     linear_elastic_plastic_isotropic_hardening,
 )
+from ._linear_elastic_viscoelastic import linear_elastic_viscoelastic
 
 __all__ = [
     "linear_elastic",
+    "linear_elastic_viscoelastic",
     "linear_elastic_plastic_isotropic_hardening",
     "LinearElasticPlasticIsotropicHardening",
 ]